Extracted WebglShader from WebglGraphicsDevice
